Rework the parsec_object_t system.
The new destructor need to return a continuation value, 0 if the chain
of destructors will contiune to be called, 1 if no other destructors
shall be called on this object. As a result, we can now recycle
parsec_objects for as long as they are correcty reinitialized twice. The
first time, when the object is stored in some internal cache (in order
to be able to properly chain it) and the second time before returning a
newly allocated object back to the user.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>